add tests to test displaying relations for enums and records that implement interfaces
update README.md

describe position of fillImport method is the ClassVoidVisitor.visit(ClassOrInterfaceDeclaration, PUml) method. Repeat here:

Imports must be filled before traversing class fields if we want to add associations, compositions, or aggregations for fields or methods, because we need access to the import declarations when checking the fields/methods.
Without this option, we could move the call to fillImports right above the addRelation call.
For now, I have not added this functionality, but will leave the call to fillImports here.