Fixed the constructor name for nested records and allowed adding relations for nested types in all TypeDeclarations.
The fillImports method calls the parseImport method, which not only parses imports but also changes the name of pUmlClass if it is nested by adding its outer class and a $. So, when the RecordDeclaration tries to create a constructor, it uses pUmlClass.getName(), and the constructor for the nested type will be something like Outer$Nested(). To avoid this, I replaced the use of pUmlClass.getName() with recordDeclaration.getNameAString(), which uses the actual name of the nested type.

This issue was not visible in the previous version because the invocation of the parseImport method occurred after creating the record's constructor.

Regarding adding relations for nested types in all declarations (classes, interfaces, records, enums, annotations), I used a supertype TypeDeclaration<?> instead of specifying the concrete type. I believe this is more appropriate, as nested types can be present inside any of the mentioned type declarations.

Added comments about the LinkedHashSet in the visit method for RecordDeclaration.

Fixed the name of the local variable parm to param.