fixed previous tests according to the provided changes
For Record3.java and Record4.java added body to the getters, because those methods did not have bodies and the parser made them abstract. But actually getters without bodies in the RecordDeclaration are compile errors.

For ClassWithConstructorOutput.txt and ClassWOConstructorOutput.txt removed <<Create>> statement and added empty lines.

Probably need to rework UmlGenerationTester to ignore empty lines it the future.